Makefile: Add update-embedded-root rule (#1301)
* Makefile: Add update-embedded-root rule

This uses the "plumbing" command to ensure the newest root has been
downloaded and verified. Then it copies the newest TUF root and the
trusted_root.json into the sources. The benefit here is that one does
not need to manually find the cache directories when an update should
be done.

This hard codes XDG_DATA_HOME and XDG_CACHE_HOME for simplicity.

We could later add a workflow that runs this on cron and files an
issue if the sources changed as a result.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* workflows: Create issue if TUF root is not up-to-date

Creates a new issue once a week if
* the embedded TUF root (or trusted_root.json) differs from the
  current one served by root-signing
* and there is no open issue with same label already

This does add a new CI-dependency (github-script) but I believe the
currently used actions do not provide the capabilities needed here.

The "embedded-root-update" label likely needs to be created by a
maintainer manually.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>