_internal/trust: Fix bug in rekor key lookup (#1350)
* _internal/trust: Fix bug in rekor key lookup

Rekor keyring can (and in future will) have multiple keys:
logs not only get sharded but once rekor-tiles is integrated in the
public good instance, there will be two writable logs for a while.

As far as I can tell all calling code is already capable of handling
the keyring.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* CHANGELOG: Mention fix for multiple rekor keys

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>