Update staging assets, refactor TUF asset lookup
Commit is large just because the test and embedded assets for staging
are updated.

* Update the embedded data in sigstore/_store
* Also update the test assets in test/assets
* refactor the embedded asset lookup: use the URL to build the
  asset dir. This means less code duplication and easier to make this
  work with non-Public Good Instance TUF repos
* Make the tuf module work with non-PGI instances: if the local TUF
  metadata is initialized out of band, tuf module just works with it.
  If a root.json is provided in _store, it is still always used to
  initialize the client

Of special note is "signing_config.v0.2.json" for production: This does
not actually exist yet in the TUF repository but I've added one in
sigstore/_store and use it as a workaround in
ClientTrustConfig.from_tuf() -- this way the code can otherwise remain
identical for both staging and prod.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>