conformance: Support --trusted-root
conformance clients are given "--trusted-root" but sigstore-python
requires "--trust-config". Build a trust config and provide that in the
conformance client script.

The conformance scipt is getting closer and closer to the point where
just tweaking argv is not really the smart thing to do... but it's still
manageable.

This revealed a new bug so one test remains "xfail".

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>