verify: Handle case of wrong Sigstore instance better
This is not foolproof but it does work when e.g. a bundle is
signed with "--staging" and verified without it:

$ sigstore verify identity ...
[16:09:18] WARNING  A certificate chain was not valid, are you using the correct Sigstore   _cli.py:1095
                    instance?
           ERROR    FAIL: sign-main-rekorv1/README.md                                       _cli.py:1098
$

In practice if TSA cert validation fails we take that as a hint that this
might be a case of wrong instance (because that's the first use case of
trusted root certs).

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>