Update embedded TUF root (#1515)
* Makefile: Update the TUF root update rules

* Include the new signingconfig file
* Add a rule for updating staging

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* Update embedded TUF root

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* Update embedded staging root

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* tests: Stop mocking TUF

The mock was useful to ensure we don't e.g. end up downloading files
multiple times but maintaining it is a bit of a hassle: the mock files
have to be kept in sync with actual staging tuf (alternatively we'd
have to create our own local tuf repository which would be another kind
of hassle).

Remove the uses of mock_staging_tuf (except for test_trust_root_tuf_offline
that ensures we do not call tuf repo when offline).

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>