cli: Add --rekor-version to sign arguments (#1471)
* cli: Add --rekor-version to sign arguments

This should not be needed... but it could be handy if
* SigningConfig already contains rekor v2
* user for some reason does not want rekor v2 entries in
  the bundle

This option only does anything if there are multiple Rekor versions
listed in SigningConfig.

The test is changed since the "ANY" selector is now considered to not be
an error if there are 0 services:
* This is not a problem since for both TSAs and tlogs we have a check
  that there is at least one service
* This improves the error message when --rekor-version is used with
  a version that is not found in signingconfig

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* README: Update help output

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* cli: Improve help output for --rekor-version

Avoid saying "default: None", mention the valid values instead.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>