tests: Try to not fail negative key tests if support is not there
We want to test the _get_key_details() with some unexpected
cryptographic material, but we don't want the test suite to fail
while collecting tests if cryptography itself does not support
these keys (since we're only testing failure cases here).

Delay the key generation code with a lambda so we can handle
cryptography exception and skip the test in these negative tests.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>